Add new files and info
The pasticho.html file was completed in terms of information. The empanadas.html file was created to continue working on it.
A new folder named "recipes" was created to contain the previously named files. A new folder was created to contain image
resources such as empanadas.jpg and pasticho.jpg.